requirements.txt
torch
torchvision
ultralytics
opencv-python-headless
matplotlib
pandas
pillow
seaborn
scikit-learn
tqdm
tensorboard
pyyaml
pytesseract
streamlit
torchmetrics
